Finished Version 1.1
Everything works now. The interface is nicer and you can mess up the Luamacros path (as far as i know. No more bugs have been found at the time of writing)

but the normal mode doesnt work for the script right now. It has the be started from the command line in No Color or No Delete mode to work!